refactor: avata 이름 변경 및 이동
avata -> avatar 이름 변경
