+++ frontend/host-app/src/components/Header/HeaderConfigAvatar.js
@@ -2,18 +2,19 @@ import {makeStyles, Icon} from "@material-ui/core";
 import Avatar from "@material-ui/core/Avatar";
 import React from "react";
 
-function HeaderConfigAvatar({onClick}) {
-	const useStyles = makeStyles({
-		headerAvatar: {
-			backgroundColor: "#FFF",
-			color: "#212529",
-			margin: "0.5rem",
-			"&:hover": {
-				color: "#FFF",
-				backgroundColor: "#69747F",
-			},
+const useStyles = makeStyles({
+	headerAvatar: {
+		backgroundColor: "#FFF",
+		color: "#212529",
+		margin: "0.5rem",
+		"&:hover": {
+			color: "#FFF",
+			backgroundColor: "#69747F",
 		},
-	});
+	},
+});
+
+function HeaderConfigAvatar({onClick}) {
 	const classes = useStyles();
 
 	return (
